数論変換 int に対応

+++ Test/NumberTheoreticTransform.test.cpp
@@ -9,10 +9,10 @@ int main() {
 	int N, M;
 	scanf("%d %d", &N, &M);
 	
-	std::vector<long long> A(N), B(M);
-	for (int i = 0; i < N; ++i) scanf("%lld", &A[i]);
-	for (int i = 0; i < M; ++i) scanf("%lld", &B[i]);
+	std::vector<int> A(N), B(M);
+	for (int i = 0; i < N; ++i) scanf("%d", &A[i]);
+	for (int i = 0; i < M; ++i) scanf("%d", &B[i]);
 	
 	auto ans = NumberTheoreticTransform<998'244'353, 3>::multiply(A, B);
-	for (int i = 0; i < N + M - 1; ++i) printf("%lld%c", ans[i], i == N + M - 1 ? '\n': ' ');
+	for (int i = 0; i < N + M - 1; ++i) printf("%d%c", ans[i], i == N + M - 1 ? '\n': ' ');
 }
