+++ Test/NumberTheoreticTransform_AnyMod.test.cpp
@@ -9,11 +9,11 @@ int main() {
 	int N, M;
 	scanf("%d %d", &N, &M);
 	
-	std::vector<long long> A(N), B(M);
-	for (int i = 0; i < N; ++i) scanf("%lld", &A[i]);
-	for (int i = 0; i < M; ++i) scanf("%lld", &B[i]);
+	std::vector<int> A(N), B(M);
+	for (int i = 0; i < N; ++i) scanf("%d", &A[i]);
+	for (int i = 0; i < M; ++i) scanf("%d", &B[i]);
 	
 	auto ans = NumberTheoreticTransform_AnyMod<1'000'000'007>::multiply(A, B);
-	for (int i = 0; i < N + M - 1; ++i) printf("%lld%c", ans[i], i == N + M - 1 ? '\n': ' ');
+	for (int i = 0; i < N + M - 1; ++i) printf("%d%c", ans[i], i == N + M - 1 ? '\n': ' ');
 	return 0;
 }
